move /external-agent/{agent}/credentials endpoint to enterprise
