+++ coderd/apidoc/docs.go
@@ -10315,7 +10315,7 @@ const docTemplate = `{
                     "application/json"
                 ],
                 "tags": [
-                    "Agents"
+                    "Enterprise"
                 ],
                 "summary": "Get workspace external agent credentials",
                 "operationId": "get-workspace-external-agent-credentials",
